Add video generation app Fixes #217
