+++ apps/video/package.json
@@ -17,7 +17,7 @@
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
     "remotion": "4.0.210",
-    "zod": "^3.23.8"
+    "zod": "3.22.3"
   },
   "devDependencies": {
     "@remotion/eslint-config": "4.0.210",
