+++ apps/video/src/Root.tsx
@@ -2,7 +2,7 @@ import "./global.css";
 
 import { loadAllLocalesAsync } from "@octocoach/i18n/src/i18n-util.async";
 import { useEffect, useState } from "react";
-import { Composition } from "remotion";
+import { Composition, continueRender, delayRender } from "remotion";
 
 import { compSchema, CourseData, MyComposition } from "./Composition";
 
@@ -26,49 +26,48 @@ const title = "AI Web App Development";
 const animatedLogo = false;
 
 export const RemotionRoot: React.FC = () => {
-  const [localesLoaded, setLocalesLoaded] = useState(false);
+  const [delayRenderId] = useState(() => delayRender());
 
   useEffect(() => {
-    void (async () => {
-      await loadAllLocalesAsync();
-      setLocalesLoaded(true);
-    })();
-  }, []);
+    if (delayRenderId) {
+      void loadAllLocalesAsync().then(() => {
+        continueRender(delayRenderId);
+      });
+    }
+  }, [delayRenderId]);
 
   return (
-    localesLoaded && (
-      <>
-        <Composition
-          id="Square"
-          component={MyComposition}
-          durationInFrames={durationInSeconds * fps}
-          fps={fps}
-          width={1920}
-          height={1920}
-          schema={compSchema}
-          defaultProps={{
-            layout: "square",
-            title,
-            animatedLogo,
-            courseData,
-          }}
-        />
-        <Composition
-          id="Portrait"
-          component={MyComposition}
-          durationInFrames={durationInSeconds * fps}
-          fps={fps}
-          width={1080}
-          height={1920}
-          schema={compSchema}
-          defaultProps={{
-            layout: "portrait",
-            title,
-            animatedLogo,
-            courseData,
-          }}
-        />
-      </>
-    )
+    <>
+      <Composition
+        id="Square"
+        component={MyComposition}
+        durationInFrames={durationInSeconds * fps}
+        fps={fps}
+        width={1920}
+        height={1920}
+        schema={compSchema}
+        defaultProps={{
+          layout: "square",
+          title,
+          animatedLogo,
+          courseData,
+        }}
+      />
+      <Composition
+        id="Portrait"
+        component={MyComposition}
+        durationInFrames={durationInSeconds * fps}
+        fps={fps}
+        width={1080}
+        height={1920}
+        schema={compSchema}
+        defaultProps={{
+          layout: "portrait",
+          title,
+          animatedLogo,
+          courseData,
+        }}
+      />
+    </>
   );
 };
